Struct sliding_windows::Window

pub struct Window<'a, T: 'a> { /* fields omitted */ }

This is the Item type of the Adaptor iterator.

Usage:

&Window<'a, T> implements into_iter(), which returns an Iterator over &T.

use sliding_windows::IterExt;
use sliding_windows::Storage;

let mut storage: Storage<u32> = Storage::new(3);
let mut windowed_iter = (0..5).sliding_windows(&mut storage);

for mut window in windowed_iter {
    // extra scope, so that later mutable borrow is possible
    {
        for x in &window {
            // work with data immutably
        }
    }

    // mutable
    let mut iter_mut = window.iter_mut();
    for x in iter_mut {
        // work with data mutably (affecting the next windows of course)
    }
}

See sliding_windows for more information.

Methods

impl<'a, T> Window<'a, T>

fn iter(&self) -> WindowIter<T>

fn iter_mut(&mut self) -> WindowIterMut<T>